requirements.txt
anthropic>=0.18.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
